+++ app/controllers/carto/api/overlay_presenter.rb
@@ -4,20 +4,25 @@ module Carto
   module Api
     class OverlayPresenter
 
-      PUBLIC_VALUES = [:id, :order, :type, :template, :options, :visualization_id]
+      PUBLIC_VALUES =  [:id, :order, :type, :template, :options, :visualization_id].freeze
+      VIZJSON_VALUES = [:type, :order, :options, :template].freeze
 
       def initialize(overlay)
         @overlay = overlay
       end
 
       def to_poro
-         public_values(@overlay)
+        values(@overlay, PUBLIC_VALUES)
+      end
+
+      def to_vizjson_poro
+        values(@overlay, VIZJSON_VALUES)
       end
 
       protected
 
-      def public_values(overlay)
-        Hash[ PUBLIC_VALUES.map { |attribute| [attribute, overlay.send(attribute)] } ]
+      def values(overlay, attribute_list)
+        Hash[attribute_list.map { |attribute| [attribute, overlay.send(attribute)] }]
       end
 
     end