+++ spec/models/visualization/copier_spec.rb
@@ -12,7 +12,6 @@ include CartoDB
 describe Visualization::Copier do
   before do
     @db = Rails::Sequel.connection
-    CartoDB::Overlay.repository = DataRepository::Backend::Sequel.new(@db, :overlays)
     Visualization.repository = DataRepository::Backend::Sequel.new(@db, :visualizations)
 
     @user = create_user
@@ -60,6 +59,20 @@ describe Visualization::Copier do
 
       copy.name.should_not be_nil
     end
+
+    it 'copies overlays' do
+      visualization = factory
+      copy          = Visualization::Copier.new(@user, visualization).copy
+
+      copy.overlays.zip(visualization.overlays).each do |overlay_copy, overlay_original|
+        overlay_copy.id.should               neq overlay_original.id
+        overlay_copy.visualization_id.should eq  copy.id
+        overlay_copy.type.should             eq  overlay_original.type
+        overlay_copy.template.should         eq  overlay_original.template
+        overlay_copy.options.should          eq  overlay_original.options
+        overlay_copy.order.should            eq  overlay_original.order
+      end
+    end
   end #copy
 
   def factory
@@ -81,4 +94,3 @@ describe Visualization::Copier do
     @user.destroy
   end
 end # Visualization::Copier
-