+++ spec/requests/api/visualizations_spec.rb
@@ -6,9 +6,6 @@ require 'uri'
 require_relative '../../spec_helper'
 require_relative '../../../app/controllers/api/json/visualizations_controller'
 require_relative '../../../services/data-repository/backend/sequel'
-require_relative '../../../app/models/visualization/migrator'
-require_relative '../../../app/models/overlay/migrator'
-
 
 def app
   CartoDB::Application.new
@@ -40,7 +37,6 @@ describe Api::Json::VisualizationsController do
     Sequel.extension(:pagination)
 
     CartoDB::Visualization.repository = DataRepository::Backend::Sequel.new(@db, :visualizations)
-    CartoDB::Overlay.repository       = DataRepository::Backend::Sequel.new(@db, :overlays)
 
     begin
       delete_user_data @user
@@ -49,7 +45,7 @@ describe Api::Json::VisualizationsController do
       raise unless exception.class.to_s == 'CartoDB::NamedMapsWrapper::HTTPResponseError'
     end
 
-    @headers = { 
+    @headers = {
       'CONTENT_TYPE'  => 'application/json',
     }
     host! 'test.localhost.lan'
@@ -104,7 +100,7 @@ describe Api::Json::VisualizationsController do
       source_visualization  = table.fetch('table_visualization')
 
       payload = { source_visualization_id: source_visualization.fetch('id') }
-      
+
       post "/api/v1/viz?api_key=#{@api_key}",
         payload.to_json, @headers
 
@@ -173,7 +169,7 @@ describe Api::Json::VisualizationsController do
         source_visualization_id:  visualization.fetch('id'),
         name:                     visualization_name
       }
-      
+
       post "/api/v1/viz?api_key=#{@api_key}",
         payload.to_json, @headers
       last_response.status.should == 200
@@ -336,7 +332,7 @@ describe Api::Json::VisualizationsController do
       source_visualization_id = table.fetch('table_visualization').fetch('id')
 
       payload = { source_visualization_id: source_visualization_id }
-      
+
       post "/api/v1/viz?api_key=#{@api_key}", payload.to_json, @headers
       response          = JSON.parse(last_response.body)
       visualization_id  = response.fetch('id')
@@ -345,7 +341,7 @@ describe Api::Json::VisualizationsController do
       last_response.status.should == 200
 
       delete "/api/v1/tables/#{table_id}?api_key=#{@api_key}", {}, @headers
-      
+
       get "/api/v1/viz/#{visualization_id}?api_key=#{@api_key}", {}, @headers
       last_response.status.should == 404
 
@@ -642,7 +638,7 @@ describe Api::Json::VisualizationsController do
     privacy = options.fetch(:privacy, 1)
 
     seed    = rand(9999)
-    payload = { 
+    payload = {
       name:         "table #{seed}",
       description:  "table #{seed} description"
     }
@@ -658,4 +654,3 @@ describe Api::Json::VisualizationsController do
     table_attributes
   end #table_factory
 end # Api::Json::VisualizationsController
-